Add files via upload
This project provides a new way to maintain the visitor entries which saved in the database, so that any time the admin can view the entries details of 
the visitor as well as the total amount collected by the entries. It also keeps the track of animals' data 
with its unique id in such a way that the data of the animals are loaded into the database. The unique 
id also used to display the loaded data. The administrator has no limitation in handling data and 
he/she has the power to allow the new user into the system. While the other officers have own 
limitation in handling data. Zoo management system will provide diplomatic way in maintain the 
zoo. Moreover, it enables the management to get reports for making quick decisions, which are not 
possible in the existing manual system